# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_import_reranker_model_multipart_v1_reranker_models_import_post": {
        "properties": {
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override linked Dataset UUID",
            "title": "Dataset Id"
          },
          "evaluation_set_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Set Id"
          },
          "external_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Source"
          },
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "model_file": {
            "contentMediaType": "application/octet-stream",
            "description": "LightGBM model.txt",
            "title": "Model File",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override RerankerModel.name",
            "title": "Name"
          },
          "prediction_set_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Set Id"
          },
          "run_json": {
            "contentMediaType": "application/octet-stream",
            "description": "Lab run.json with metrics + provenance",
            "title": "Run Json",
            "type": "string"
          },
          "spec_yaml": {
            "contentMediaType": "application/octet-stream",
            "description": "ExperimentSpec spec.yaml",
            "title": "Spec Yaml",
            "type": "string"
          }
        },
        "required": [
          "model_file",
          "spec_yaml",
          "run_json"
        ],
        "title": "Body_import_reranker_model_multipart_v1_reranker_models_import_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImportRerankerByReferenceRequest": {
        "additionalProperties": false,
        "description": "Body for ``POST /reranker-models/import-by-reference``.\n\nUse this when the lab has already uploaded ``model.txt`` to MinIO\nunder its own key and just needs PROTEA to register the URI.",
        "example": {
          "artifact_uri": "s3://protea-rerankers/runs/r1/model.txt",
          "dataset_id": "00000000-0000-0000-0000-000000000003",
          "evaluation_set_id": "00000000-0000-0000-0000-000000000005",
          "external_source": "protea-reranker-lab@cec8ccd",
          "force": false,
          "name": "r1-k5-bench-v1",
          "prediction_set_id": "00000000-0000-0000-0000-000000000004",
          "run": {
            "feature_schema_sha": "ab12cd34ef56",
            "metrics": {
              "fmax": 0.5427
            },
            "run_id": "r1"
          },
          "spec_yaml": "# ExperimentSpec contents\nname: r1\nfeature_families: [embedding, alignment]\n"
        },
        "properties": {
          "artifact_uri": {
            "description": "Storage URI of the booster blob (``s3://\u2026/model.txt`` or ``file://\u2026/model.txt``). PROTEA registers this verbatim without re-reading or copying the artifact.",
            "minLength": 1,
            "title": "Artifact Uri",
            "type": "string"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "UUID of the source ``Dataset`` row this booster was trained on. Used to track the lab \u2192 PROTEA lineage.",
            "title": "Dataset Id"
          },
          "evaluation_set_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional UUID of an ``EvaluationSet`` the booster was tuned against (NK / LK / PK delta).",
            "title": "Evaluation Set Id"
          },
          "external_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Free-form provenance tag (e.g. ``protea-reranker-lab@<git-sha>``) for boosters trained outside PROTEA.",
            "title": "External Source"
          },
          "force": {
            "default": false,
            "description": "When true, allow overwriting an existing row with the same ``name``. Default false rejects duplicates with 409.",
            "title": "Force",
            "type": "boolean"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override row name. When omitted, falls back to the ``run.run_id`` from the run.json payload.",
            "title": "Name"
          },
          "prediction_set_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional UUID of a ``PredictionSet`` whose vote pattern the booster targets; nullable so abstract boosters can be registered without binding to a specific run.",
            "title": "Prediction Set Id"
          },
          "run": {
            "additionalProperties": true,
            "description": "Parsed run.json",
            "title": "Run",
            "type": "object"
          },
          "spec_yaml": {
            "description": "Full spec.yaml text",
            "title": "Spec Yaml",
            "type": "string"
          }
        },
        "required": [
          "artifact_uri",
          "spec_yaml",
          "run"
        ],
        "title": "ImportRerankerByReferenceRequest",
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "Pydantic model for an RFC 7807 problem-details payload.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Human-readable explanation of this specific occurrence, distinct from the generic ``title``.",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "URI of the specific occurrence (typically the request path); helps correlate with logs.",
            "title": "Instance"
          },
          "status": {
            "description": "HTTP status code repeated in the body so the payload is self-contained for clients that drop the status line.",
            "maximum": 599,
            "minimum": 100,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, stable summary of the problem.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI reference identifying the problem class. Relative URIs resolve under the API host; use ``about:blank`` for generic.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "contact@protea.example.org",
      "name": "PROTEA Team"
    },
    "description": "**PROTEA** \u2014 Protein Representation and Ontology-Term Enrichment Analysis.\n\nManages the full pipeline from UniProt sequence ingestion through GPU embedding computation (ESM-2, ESM3c, T5) to KNN-based GO term prediction.\n\nAll long-running operations are queued via RabbitMQ and tracked as `Job` rows with a full event audit trail. Use `GET /jobs/{id}/events` to stream real-time progress.",
    "title": "PROTEA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/reranker-models/import": {
      "post": {
        "description": "Upload a trained booster and register a ``RerankerModel`` row.\n\nThe three files (``model.txt``, ``spec.yaml``, ``run.json``) mirror\nthe artefacts produced by ``protea-reranker-lab`` under\n``runs/<name>/``. Wire format unchanged: the FastAPI deps expose\nevery File/Form field as a discrete multipart part.",
        "operationId": "import_reranker_model_multipart_v1_reranker_models_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_reranker_model_multipart_v1_reranker_models_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Import Reranker Model Multipart V1 Reranker Models Import Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import a lab-trained booster",
        "tags": [
          "reranker-models"
        ]
      }
    },
    "/v1/reranker-models/import-by-reference": {
      "post": {
        "description": "Register a ``RerankerModel`` whose booster is already in MinIO.\n\nThe lab uploads the booster directly (faster, no double-hop) and\nPOSTs the URI + run.json + spec.yaml here. Server does not re-read\nthe artifact; it trusts the URI.",
        "operationId": "import_reranker_model_by_reference_v1_reranker_models_import_by_reference_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRerankerByReferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Import Reranker Model By Reference V1 Reranker Models Import By Reference Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register a booster already in the artifact store",
        "tags": [
          "reranker-models"
        ]
      }
    },
...
  },
...
}